move: コンポーネントのディレクトリ構成の変更
