+++ frontend/src/app/(not_login)/page.tsx
@@ -1,6 +1,6 @@
 import CheckRootPage from "@/components/auth/check_root_page";
-import { LogoutButton } from "@/components/common/button_logout";
-import { GoogleLoginButton } from "@/components/common/button_login_google";
+import { LogoutButton } from "@/components/common/button/logout";
+import { GoogleLoginButton } from "@/components/common/button/login_google";
 
 export default function RootPage() {
   return (
