+++ frontend/src/components/common/button/new_post.tsx
